modules: uds: modify sample to successfully recv
Previously the module sample only compiled but
did not demonstrate any UDS functionality. This
version is based largely on the esp32 example
in driftregion/iso14229.

The implementation uses k_msgq and filter
msgq callbacks to queue the CAN frames, and uses
the app main loop to shovel the CAN frames
to the lib isotp interface. The app main loop also
triggers the library UDS Server polling.

This includes a test script for sending a
ECUReset message over can0 that this version
of the module sample can receive and act on.

Signed-off-by: Oskar Lundstrom <oslundstrom@gmail.com>