modules: uds: init module for external UDS lib
Zephyr does not have a implementation of UDSonCAN,
driftregion/iso14229 has implemented UDSonCAN partly
with embedded isotp library. This implementation
includes the library as a default module.

The implementation will include an associated
sample that is based off of the isotp sample.

The module is placed under lib and not under
CAN since the specification is viable for other
links than CAN.

Status of this code is that it compiles (including
the external library), but does
not do anything.

Signed-off-by: Oskar Lundstrom <oslundstrom@gmail.com>